C++ snippets to handle serialization

+++ crates/target_cpp/src/cpp_snippets/enum_to_json.cpp
@@ -1,5 +1,5 @@
 
   ExpType<void> serialize(Writer::Serializer& serializer, const $FULL_NAME$ value) {
     using Enum = $FULL_NAME$;
-    return serializer.write_str(Common<Enum>::entries[size_t(value)]);
+    return serializer.write_str(Common<Enum>::entries[int(value)]);
   }
